#11 fixed bots.py

+++ src/models/bet.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class Bet(BaseModel):
     id: int
-    bet_value: int
+    bet_value: float
     bot_id: int
     game_id: str
     end_time: int
